# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@apihub.com
    name: API Support
    url: https://github.com/ultimatum/apihub_go
  description: A comprehensive API hub with multiple API categories for learning and
    development
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
  termsOfService: http://swagger.io/terms/
  title: FreeAPI - API Hub
  version: "1.0"
paths:
  /kitchen-sink/cookies/delete:
    get:
      parameters:
      - description: Cookie Key
        in: query
        name: key
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete a cookie
      tags:
      - Kitchen Sink
...
  /kitchen-sink/cookies/set:
    get:
      parameters:
      - description: Cookie Key
        in: query
        name: key
        required: true
        type: string
      - description: Cookie Value
        in: query
        name: value
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Set a cookie
      tags:
      - Kitchen Sink
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"